PLFENG-1968: fix update source UpdateSourcesMojo

+++ plugin/src/main/java/org/exoplatform/crowdin/mojo/UpdateSourcesMojo.java
@@ -26,6 +26,7 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
@@ -41,7 +42,9 @@ import org.exoplatform.crowdin.model.CrowdinFile.Type;
 import org.exoplatform.crowdin.model.CrowdinFileFactory;
 import org.exoplatform.crowdin.model.CrowdinTranslation;
 import org.exoplatform.crowdin.model.SourcesRepository;
+import org.exoplatform.crowdin.utils.IOSResouceBundleFileUtils;
 import org.exoplatform.crowdin.utils.PropsToXML;
+import org.exoplatform.crowdin.utils.XMLResourceBundleUtils;
 
 /**
  * Update projects sources from crowdin translations
@@ -183,9 +186,32 @@ public class UpdateSourcesMojo extends AbstractCrowdinMojo {
           }
           String key = zipentryName.substring(zipentryName.indexOf(cp) + cp.length() + 1);
           String value = currentProj.getProperty(key);
+          
           if (value == null) {
             zipentry = zipinputstream.getNextEntry();
             continue;
+          }
+          
+          /**
+           * if android, don't save to default master folder but save to
+           * "values-language" folder (for example) res/values/strings.xml > res/values-fr/strings.xml
+           */
+          if (zipentryName.contains("android")) {
+            if (!locale.contains("en")) {
+              String localizable = CrowdinTranslation.encodeAndroidLocale(locale);
+              value = value.replace("res/values/", "res/values-" + localizable + "/");
+            }
+          }
+          /**
+           * if iOS, don't save to default master folder but save to
+           * "language.proj" folder (for example)
+           * /Resources/en.lproj/Localizable.string > /Resources/fr.lproj/Localizable.string
+           */
+          else if (zipentryName.contains("ios")) {
+            if (!locale.contains("en")) {
+              String localizable = CrowdinTranslation.encodeIOSLocale(locale);
+              value = value.replace("en.lproj", localizable + ".lproj");
+            }
           }
           zipentryName = zipentryName.substring(0, zipentryName.indexOf(proj) + proj.length());
 
@@ -210,7 +236,16 @@ public class UpdateSourcesMojo extends AbstractCrowdinMojo {
               fileName = name + "_" + lang + extension;
             }
 
-          } else {
+          }
+          //if android, don't change xml to properties
+          else if (zipentryName.contains("android") ){
+            fileName = name + ".xml";
+          }
+          //if iOS
+          else if(zipentryName.contains("ios") ){
+            fileName = name + extension;
+          }
+          else {
             fileName = name + "_" + lang + extension;
           }
 
@@ -227,20 +262,92 @@ public class UpdateSourcesMojo extends AbstractCrowdinMojo {
           boolean isXML = (entryName.indexOf(".xml") > 0);
 
           if (isXML) {
-            // create the temporary properties file to be used for PropsToXML (use the file in Crowdin zip)
-            entryName = entryName.replaceAll(".xml", ".properties");
+            //if is Android resouce bundle
+            if(zipentryName.contains("mobile") && zipentryName.contains("android")){
+              String resourceTranslationFilePath  = parentDir + name + extension;
+              String localizable = CrowdinTranslation.encodeAndroidLocale(locale);
+              String masterFilePath = resourceTranslationFilePath.replaceAll("-" + localizable, "");
+
+              //create temporary file to persists zipinputstream
+              int n;
+              FileOutputStream fileoutputstream;
+              fileoutputstream = new FileOutputStream(resourceTranslationFilePath+".ziptempo");
+              while ((n = zipinputstream.read(buf, 0, 1024)) > -1) {
+                fileoutputstream.write(buf, 0, n);
+              }
+              fileoutputstream.close();
+              String crowdinFilePath = resourceTranslationFilePath + ".ziptempo";
+              FileInputStream input = new FileInputStream(crowdinFilePath);
+              XMLResourceBundleUtils.setLog(getLog());
+              XMLResourceBundleUtils.injectTranslation(input, resourceTranslationFilePath, masterFilePath);
+              
+              //delete ziptempo file
+              try{
+                File file = new File(crowdinFilePath);
+                if(file.delete()){
+                  if(getLog().isDebugEnabled())
+                    getLog().debug(file.getName() + " is deleted!");
+                }else{
+                  if(getLog().isDebugEnabled())
+                    getLog().debug("Delete operation is failed.");
+                }
+              }catch(Exception e){
+                getLog().error(e);
+              }
+            }
+            else{
+              // create the temporary properties file to be used for PropsToXML (use the file in Crowdin zip) 
+              //if not in mobile project, convert xml to properties
+              if (!zipentryName.contains("mobile")) {
+                entryName = entryName.replaceAll(".xml", ".properties");
+              }
+              
+              int n;
+              FileOutputStream fileoutputstream;
+              fileoutputstream = new FileOutputStream(entryName);
+              while ((n = zipinputstream.read(buf, 0, 1024)) > -1) {
+                fileoutputstream.write(buf, 0, n);
+              }
+              fileoutputstream.close();
+  
+              File propertiesFile = new File(entryName);     
+              
+              // don't convert to ascii in mobile project
+              if (!zipentryName.contains("mobile")) {
+                PropsToXML.execShellCommand("native2ascii -encoding UTF8 " + propertiesFile.getPath()
+                    + " " + propertiesFile.getPath());
+              }
+  
+              PropsToXML.parse(propertiesFile.getPath(), resourceBundleType);
+              propertiesFile.delete();
+            }
+          }
+          
+          // when project is iOS
+          else if(zipentryName.contains("ios")){
+            
+            String localFile = parentDir + name + extension;
+            String localizable = CrowdinTranslation.encodeIOSLocale(locale);
+            String masterFile = localFile.replace(localizable + ".lproj", "en.lproj");   
+            String resourceTranslationFilePath = localFile;            
+            //Write tempo zipinputstream
             int n;
             FileOutputStream fileoutputstream;
-            fileoutputstream = new FileOutputStream(entryName);
+            fileoutputstream = new FileOutputStream(resourceTranslationFilePath+".ziptempo");
             while ((n = zipinputstream.read(buf, 0, 1024)) > -1) {
               fileoutputstream.write(buf, 0, n);
             }
             fileoutputstream.close();
-
-            File propertiesFile = new File(entryName);
-            PropsToXML.execShellCommand("native2ascii -encoding UTF8 " + propertiesFile.getPath() + " " + propertiesFile.getPath());
-            PropsToXML.parse(propertiesFile.getPath(), resourceBundleType);
-            propertiesFile.delete();
+            
+            String crowdinFilePath = resourceTranslationFilePath + ".ziptempo";
+              
+              //master file code base EN
+            String resourceMasterFilePath = masterFile;
+              
+            //translation file code base LANGUAGE
+            IOSResouceBundleFileUtils.setLog(getLog());
+            IOSResouceBundleFileUtils.injectTranslation(crowdinFilePath, resourceMasterFilePath, resourceTranslationFilePath);
+          
           } else {
             // identify the master properties file
             String masterFile = parentDir + name + extension;
